+++ examples/nextjs/pages/guarded-ssr.tsx
@@ -4,10 +4,9 @@ import { Container, Title } from '@mantine/core'
 import { getNhostSession, NhostSession, useAccessToken } from '@nhost/nextjs'
 
 import { authProtected } from '../components/protected-route'
-import { BACKEND_URL } from '../helpers'
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
-  const nhostSession = await getNhostSession(BACKEND_URL, context)
+  const nhostSession = await getNhostSession({ subdomain: 'localhost' }, context)
   return {
     props: {
       nhostSession