+++ examples/nextjs/pages/public-ssr.tsx
@@ -3,10 +3,8 @@ import { GetServerSideProps } from 'next'
 import { Container, Title } from '@mantine/core'
 import { getNhostSession, NhostSession, useAccessToken, useAuthenticated } from '@nhost/nextjs'
 
-import { BACKEND_URL } from '../helpers'
-
 export const getServerSideProps: GetServerSideProps = async (context) => {
-  const nhostSession = await getNhostSession(BACKEND_URL, context)
+  const nhostSession = await getNhostSession({ subdomain: 'localhost' }, context)
   return {
     props: {
       nhostSession