+++ integrations/apollo/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @nhost/apollo
 
+## 5.1.1
+
+### Patch Changes
+
+- @nhost/nhost-js@2.1.1
+
+## 5.1.0
+
+### Patch Changes
+
+- Updated dependencies [bfb4c1a6]
+  - @nhost/nhost-js@2.1.0
+
 ## 5.0.9
 
 ### Patch Changes