+++ integrations/react-urql/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @nhost/react-urql
 
+## 2.0.11
+
+### Patch Changes
+
+- @nhost/react@2.0.11
+
+## 2.0.10
+
+### Patch Changes
+
+- @nhost/react@2.0.10
+
 ## 2.0.9
 
 ### Patch Changes