+++ packages/react/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @nhost/react
 
+## 2.0.11
+
+### Patch Changes
+
+- @nhost/nhost-js@2.1.1
+
+## 2.0.10
+
+### Patch Changes
+
+- Updated dependencies [bfb4c1a6]
+  - @nhost/nhost-js@2.1.0
+
 ## 2.0.9
 
 ### Patch Changes